requirements.txt
streamlit
QuantLib-Python
yfinance
plotly
pandas
numpy
scipy